pyproject.toml
[tool.poetry]
name = "render-native"
version = "0.1.0"
description = "Learning poetry and configuration settings of FastAPI"
authors = ["Robin Michael <Siliconrob@siliconheaven.net>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
stamina = "^24.2.0"
httpx = "^0.27.0"
pydantic-settings = "^2.4.0"
pydapper = { version = "^0.10.0", python = "^3.12" }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
